Maven build updated
I was hesitant to push this, as it changes some things and I was
wondering whether the instructor's setup is different (as their
instructions for setting up maven were either pisspoor or for  a
different setup then I had). However if there are problems I can just
revert this commit.